# Fix infinite loop in Get